+++ SpaceRTS.sln
@@ -8,6 +8,11 @@ EndProject
 Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "ModelWriter", "ModelWriter\ModelWriter.vcxproj", "{15ED95AC-C839-4C21-B6AC-1F9A114383A7}"
 EndProject
 Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "Modeller", "Modeller\Modeller.vcxproj", "{3AF91AFC-07F6-4FF7-A972-E9DA31FFD317}"
+	ProjectSection(ProjectDependencies) = postProject
+		{50F2C82D-817B-43CD-A273-64F45320A848} = {50F2C82D-817B-43CD-A273-64F45320A848}
+	EndProjectSection
+EndProject
+Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "Engine", "Engine\Engine.vcxproj", "{50F2C82D-817B-43CD-A273-64F45320A848}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -39,6 +44,14 @@ Global
 		{3AF91AFC-07F6-4FF7-A972-E9DA31FFD317}.Release|Win32.Build.0 = Release|Win32
 		{3AF91AFC-07F6-4FF7-A972-E9DA31FFD317}.Release|x64.ActiveCfg = Release|x64
 		{3AF91AFC-07F6-4FF7-A972-E9DA31FFD317}.Release|x64.Build.0 = Release|x64
+		{50F2C82D-817B-43CD-A273-64F45320A848}.Debug|Win32.ActiveCfg = Debug|Win32
+		{50F2C82D-817B-43CD-A273-64F45320A848}.Debug|Win32.Build.0 = Debug|Win32
+		{50F2C82D-817B-43CD-A273-64F45320A848}.Debug|x64.ActiveCfg = Debug|x64
+		{50F2C82D-817B-43CD-A273-64F45320A848}.Debug|x64.Build.0 = Debug|x64
+		{50F2C82D-817B-43CD-A273-64F45320A848}.Release|Win32.ActiveCfg = Release|Win32
+		{50F2C82D-817B-43CD-A273-64F45320A848}.Release|Win32.Build.0 = Release|Win32
+		{50F2C82D-817B-43CD-A273-64F45320A848}.Release|x64.ActiveCfg = Release|x64
+		{50F2C82D-817B-43CD-A273-64F45320A848}.Release|x64.Build.0 = Release|x64
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
